refactoring to components
