+++ src/Components/Game/index.scss
@@ -12,9 +12,7 @@
   color: white;
   border-radius: 3px;
 }
-.item--timer {
-  margin-right: 15px;
-}
+
 .item__strong {
   color: yellow;
   font-weight: bold;
@@ -26,40 +24,6 @@
 .container__body {
   margin-top: 10px;
   margin-bottom: 10px;
-}
-.mole-home {
-  display: inline-block;
-  padding: 5px;
-  margin: 5px;
-  background-image: url("../../Images/frame.jpg");
-  background-repeat: no-repeat;
-  background-size: 100% 100%;
-  width: 65px;
-  height: 65px;
-}
-.mole {
-  width: 80%;
-  height: 80%;
-  margin: 5px;
-}
-.mole__on {
-  background-image: url("../../Images/mole.png");
-  background-repeat: no-repeat;
-  background-size: 100% 100%;
-  cursor: pointer;
-  animation: fadein 1s;
-}
-@keyframes fadein {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-
-hr {
-  visibility: hidden;
 }
 
 /* footer */
