+++ src/Components/Timer/index.scss
@@ -1,0 +1,3 @@
+.item--timer {
+  margin-right: 15px;
+}
