+++ src/Utils/constants.js
@@ -1,0 +1,18 @@
+const maxSeconds = 60000;
+const showSeconds = {
+  min: 1000,
+  max: 3000
+};
+const hideSeconds = 500;
+const maxMoleCnt = 5;
+const MoleCnt = 24;
+const initialMoles = Array(MoleCnt).fill(false);
+
+export {
+  maxSeconds,
+  showSeconds,
+  hideSeconds,
+  maxMoleCnt,
+  MoleCnt,
+  initialMoles
+};
